Cleaned up code and added Java Docs

+++ src/main/java/me/feinrasur/utils/gui/old/GuiOld.java
@@ -13,40 +13,38 @@ import org.bukkit.inventory.meta.ItemMeta;
 import java.util.HashMap;
 import java.util.List;
 
-public abstract class GuiOld
-{
+public abstract class GuiOld {
 
     Inventory inventory;
     HashMap<Integer, ClickEvent> events = new HashMap<>();
 
     /**
      * size is 1 indexed
+     *
      * @param size Sets the size of the GUI inventory
      * @param name Sets the name of the GUI inventory
      */
     @SuppressWarnings("deprecation")
-    public GuiOld(Integer size, String name)
-    {
+    public GuiOld(Integer size, String name) {
         inventory = Bukkit.createInventory(null, size, name);
     }
 
     /**
      * @param player Player to open GUI for
      */
-    public void open(Player player)
-    {
+    public void open(Player player) {
         player.openInventory(inventory);
         GUIListenerOld.manager.setCurrentGui(player, this);
     }
 
     /**
      * Slot is normally indexed
-     * @param itemStack ItemStack to add in to the inventory
-     * @param slot Slot where ItemStack should be
+     *
+     * @param itemStack  ItemStack to add in to the inventory
+     * @param slot       Slot where ItemStack should be
      * @param clickEvent ClickEvent for the ItemStack
      */
-    public void setItem(ItemStack itemStack, Integer slot, ClickEvent clickEvent)
-    {
+    public void setItem(ItemStack itemStack, Integer slot, ClickEvent clickEvent) {
         inventory.setItem(slot, itemStack);
         events.put(slot, clickEvent);
     }
@@ -63,22 +61,21 @@ public abstract class GuiOld
      * @param item item ItemStack to put in the GUI Inventory
      * @return slot the item got put in | Returns -1 if no slot is available
      */
-    public Integer add(ItemStack item)
-    {
+    public Integer add(ItemStack item) {
         Integer freeSlot = inventory.firstEmpty();
-        if (freeSlot.equals((int) -1)) return null;
+        if (freeSlot.equals(-1)) return null;
         inventory.setItem(freeSlot, item);
         return freeSlot;
     }
 
     /**
-     * @param item ItemStack to put in the GUI Inventory
+     * @param item       ItemStack to put in the GUI Inventory
      * @param clickEvent ClickEvent to set for the ItemStack
      * @return slot the item got put in | Returns -1 if no slot is available
      */
     public Integer add(ItemStack item, ClickEvent clickEvent) {
         Integer freeSlot = inventory.firstEmpty();
-        if (freeSlot.equals((int) -1)) return null;
+        if (freeSlot.equals(-1)) return null;
         inventory.setItem(freeSlot, item);
         createClickEvent(freeSlot, clickEvent);
         return freeSlot;
@@ -87,9 +84,8 @@ public abstract class GuiOld
     /**
      * @param itemList Adds List of ItemStacks in the GUI inventory without ClickEvent
      */
-    public void add(List<ItemStack> itemList)
-    {
-        for (ItemStack i: itemList) {
+    public void add(List<ItemStack> itemList) {
+        for (ItemStack i : itemList) {
             inventory.addItem(i);
         }
     }
@@ -98,8 +94,7 @@ public abstract class GuiOld
     /**
      * Create the ClickEvent
      */
-    public interface ClickEvent
-    {
+    public interface ClickEvent {
         void run(InventoryClickEvent event);
     }
 
@@ -107,13 +102,12 @@ public abstract class GuiOld
      * @param slot GUI Inventory slot
      * @return ClickEvent assigned to slot
      */
-    public ClickEvent getAction(Integer slot)
-    {
+    public ClickEvent getAction(Integer slot) {
         return events.get(slot);
     }
 
     /**
-     * @param slot Set the slot for the ClickEvent
+     * @param slot  Set the slot for the ClickEvent
      * @param event Set the ClickEvent for the slot
      */
     public void createClickEvent(Integer slot, ClickEvent event) {
